Api commits (#10)
* Network API

Deleted Unnecessary Files and reworked the Network API files to fulfill the requirements for the checkpoint, by adding the generalization for the files and the required interfaces.

* Process API

Deleted Unnecessary Files and reworked the Process API files to fulfill the requirements for the checkpoint, by adding the generalization for the files and the required interfaces. Removed the unnecessary code in the ProcessAPI AND ProcessAPIPrototype files

* ConceptualAPI

Deleted Unnecessary Files and reworked the Conceptual API files to fulfill the requirements for the checkpoint, by adding the generalization for the files and the required interfaces. Removed the unnecessary code in the ConceptualAPI AND ConceptualAPIPrototype files

* Added import

Added the annotation import to see if this will pass the checkpoint

* Changed file hierarchy for the api files

Added new file structure to hopefully fulfill the checkpoint

* Reworked File Heirarchy

Added a main.java package in which the project annotations and api files are stored in

* Moved Annotations

Moved annotations to project.annotations

* Moved annotations back into main folder

* Changed File directory for annotations

* Fixes to prototypes

Fixed the prototype classes by removing unnecessary files

* Fixed capitalizaton in ComputationAPI

* Fixed Semicolon in UserComputePrototype