Added Integration Test and Updated build.gradle
Added the integration test to verify the correct integration of the various components of the API by inputting a preset list of Integers with no specified delimiter, this will result in a fail since the engine hasn't been established but  will be returned correctly. Also added the exclusion of the tests from the build.gradle file as the tests will give an error.