Compute engine integration (#22)
* Added Implementations

Added the implementation classes in order to initialize the API and make sure everything works by returning default or null values

* Added Smoke Tests for APIs

Added the smoke test using Mockito and Jupiter to mock any parameters and dependencies required in order to verify the test is in fact calling from the API correctly.
Added:
-TestComputationAPI
-TestStorageComputeAPI
-TestUserComputeAPI

* Added InMemory infrastructure tests

Created a test implementation of the input/output configuration interfaces, in this case utilizing a List as the object for the test in order to verify that the input and output is running correctly.

* Added Integration Test and Updated build.gradle

Added the integration test to verify the correct integration of the various components of the API by inputting a preset list of Integers with no specified delimiter, this will result in a fail since the engine hasn't been established but  will be returned correctly. Also added the exclusion of the tests from the build.gradle file as the tests will give an error.

* Fixed Checkstyle errors

Changed all wildcard imports to proper imports and changed package name from JUnitTestFiles to junittests

* Added missing semicolon

Added missing semicolon to import

* Fixed another wildcard import

Fixed another wildcard import to a proper import.

* Fixed various bugs

-Fixed misspelled import for Jupiter in ComputeIntegrationTest
-Added missing imports such as assertFalse and assertTrue from Jupiter in ComputeIntegrationTest
-Added DataBatch import from api.storage, fixed lowercase to uppercase in databatch, changed return value to void for the test.

* Changed input from DataBatch to List<Integers>

Was not working with DataBatch so I have changed it to List<Integers>

* Fixed misspelling

Changed from Integers to Integer

* added assertFalse import

Changed the return values and reinstated the DataBatch object and added the assertFalse import

* Added Static modifiers

Added Static modifiers

* Changed boolean again

* changed modifiers again

* Changed DataBatch to List<Integer>

Changed DataBatch to List<Integer> in StorageComputeImpl.java

* Changed DataBatch to List<Integer>

Changed DataBatch to List<Integer> in TestStorageComputeAPI

* changed DataBatch to List<Integer>

changed DataBatch to List<Integer> in StorageComputeAPI

* changed modifiers

* Excluded Integration test

Excluded Integration test using a tag and exclude method in build.gradle

* Updated Prototypes and API parameters

Added parameters for ReadData and ExecuteJob methods that will read for the Data Source of what will be read or executed.

* Updated readData parameter

* Changed parameters again

* Changed Parameters #3

* Used mockito to mock the dataBatch parameter for readData

* Updated parameter for readData in InMemoryStorageComputeAPI

* Removed Test Annoation from Checkpoint 1

Initially worked but due to Checkpoint 1 test requirements does not continue to work

* Moved APIs into upper folder in /src

removed the main.api.java sub folders to see if the checkpoint tests will be able to detect the locations of the files

* removed unnecessary method

* Implementation of Compute Engine

*Added the computation component which will read the the input and run it through the computation to then return a string of letters translated from the Caesar Cipher.

*The coordination component will take in an input source, output source and potential delimiter to submit a job to request.

*Updated the data storage component to read and write data from the other components

* Updated DataBatch dependencies

*Deleted DataBatch as it didn't actually serve a purpose within the code as the input and output could just be directly passed without DataBatch.

*Updated every file that required DataBatch and replaced them with DataSource and DataDestination instead which fixes all the previous problems I had.

* Added Tests

*Added new test in ComputeEngineIntegrationTest named testEndToEndWithEmptyInput which will test what happens if the input is empty.

*Added Checkpoint 4 Tests

*Added the implementations of all 3 APIs and added the computation test in the ManualTestingFramework

* fixed try catch statements that are required for newBufferedReader/Writer

Fixed empty catch statements and missing brackets

* Fixed more try/catch statements and line breaks

fixed the line breaks in the try catch statement as well as they destination check

* Moved bracket again

Moved the bracket back to the same line to fix checkstyle error

* Empty Collections and mockDestination test

removed empty collection in ComputationImpl and added the mockDestination due to removing DataBatch

* Updated TestStorageComputeAPI to fail

Added signatures to fail the test on purpose

* Update TestStorageComputeAPI.java

* Update TestStorageComputeAPI.java

* added empty list if pointer is null

* Added more effective method to map integers

Added a method that will write all of the data as a single comma-separated line