Total main overhaul
* Added Implementations

Added the implementation classes in order to initialize the API and make sure everything works by returning default or null values

* Added Smoke Tests for APIs

Added the smoke test using Mockito and Jupiter to mock any parameters and dependencies required in order to verify the test is in fact calling from the API correctly.
Added:
-TestComputationAPI
-TestStorageComputeAPI
-TestUserComputeAPI

* Added InMemory infrastructure tests

Created a test implementation of the input/output configuration interfaces, in this case utilizing a List as the object for the test in order to verify that the input and output is running correctly.

* Added Integration Test and Updated build.gradle

Added the integration test to verify the correct integration of the various components of the API by inputting a preset list of Integers with no specified delimiter, this will result in a fail since the engine hasn't been established but  will be returned correctly. Also added the exclusion of the tests from the build.gradle file as the tests will give an error.

* Fixed Checkstyle errors

Changed all wildcard imports to proper imports and changed package name from JUnitTestFiles to junittests

* Added missing semicolon

Added missing semicolon to import

* Fixed another wildcard import

Fixed another wildcard import to a proper import.

* Fixed various bugs

-Fixed misspelled import for Jupiter in ComputeIntegrationTest
-Added missing imports such as assertFalse and assertTrue from Jupiter in ComputeIntegrationTest
-Added DataBatch import from api.storage, fixed lowercase to uppercase in databatch, changed return value to void for the test.

* Changed input from DataBatch to List<Integers>

Was not working with DataBatch so I have changed it to List<Integers>

* Fixed misspelling

Changed from Integers to Integer

* added assertFalse import

Changed the return values and reinstated the DataBatch object and added the assertFalse import

* Added Static modifiers

Added Static modifiers

* Changed boolean again

* changed modifiers again

* Changed DataBatch to List<Integer>

Changed DataBatch to List<Integer> in StorageComputeImpl.java

* Changed DataBatch to List<Integer>

Changed DataBatch to List<Integer> in TestStorageComputeAPI

* changed DataBatch to List<Integer>

changed DataBatch to List<Integer> in StorageComputeAPI

* changed modifiers

* Excluded Integration test

Excluded Integration test using a tag and exclude method in build.gradle

* Updated Prototypes and API parameters

Added parameters for ReadData and ExecuteJob methods that will read for the Data Source of what will be read or executed.

* Updated readData parameter

* Changed parameters again

* Changed Parameters #3

* Used mockito to mock the dataBatch parameter for readData

* Updated parameter for readData in InMemoryStorageComputeAPI

* Removed Test Annoation from Checkpoint 1

Initially worked but due to Checkpoint 1 test requirements does not continue to work

* Moved APIs into upper folder in /src

removed the main.api.java sub folders to see if the checkpoint tests will be able to detect the locations of the files

* removed unnecessary method

* Complete Overhaul of my code

Completely reworked how the APIs communicate between each other and pass data, which will hopefully result in an actually functional piece of code

* Fixed implementation in ManualTesting

StorageComputeImpl no longer requires ComputationAPI

* Complete Code Overhaul

Nothing worked in the original build, so I decided to fix EVERYTHING
*The program starts by running the SubmitJob method in UserComputeImpl
*The SubmitJob takes in an DataSource to retrieve the file with the data, a DataDestination to write the computed data to and an optional delimiter .
*Submit Job takes in the data from the source through StorageComputeAPI's readData which formats each entry as a individual number in an integer list and stores it in a local inputData variable.
*inputData is computed through ComputationAPI's compute method which takes every entry, translates it into a letter, then applies the Caesar cipher to decode the actual letter. After all entries have been decoded, returns all letters into a String with no delimiters.
*Finally, submitJob will take the returned string from the compute method and add it to the writeData method along with the dataDestination and a custom delimiter if there was one.
*writeData will format the string of letters into another string that is separated by the delimiter and written into a file specified at the location of dataDestination.

*The tests now check for valid inputs, exceptions, error handling

* updated import style

Updated Assertions import from * to actual statements